# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Micrositio Iris Backend API",
    "description": "API REST para el Micrositio Iris.\n\nTodas las respuestas usan formato unificado:\n  EXITO:   { success: true, data: <payload>, message?: string, total?: number, page?: number, limit?: number, totalPages?: number }\n  ERROR:   { success: false, message: string, errors?: [...] }\n\nAutenticacion JWT Bearer token: Authorization: Bearer <token>\nRoles:\n  - VISITANTE: puede ver solo sus proyectos asignados (errores 403 si intenta modificar)\n  - EDITOR: acceso total a lectura y escritura\n\nEndpoints de LECTURA (GET): requieren BearerAuth (cualquier token valido).\nEndpoints de ESCRITURA (POST/PATCH/PUT/DELETE): requieren BearerAuth + SoloEditores (token de usuario EDITOR).\n\nErrores HTTP estandar:\n  - 400: Parametro invalido o mal formado\n  - 401: Token JWT faltante, invalido o expirado\n  - 403: Rol insuficiente (VISITANTE intentando escribir O sin acceso al proyecto)\n  - 404: Recurso no encontrado (ID invalido, entidad eliminada)\n  - 422: Error de validacion de datos (Zod)\n  - 500: Error interno del servidor\n\nFlujo tipico para agentes:\n  1. POST /auth/login con { alias, password } para obtener token EDITOR\n  2. Usar el token en Authorization: Bearer <token> para todas las peticiones\n  3. GET para leer recursos, POST/PATCH/DELETE para modificarlos\n  4. Las respuestas SIEMPRE vienen envueltas en { success, data }",
    "version": "1.0.0"
  },
  "servers": [
    {
...
    {
...
  "paths": {
...
    "/proyectos/usuario/{id_usuario}": {
      "get": {
        "tags": [
          "Proyectos"
        ],
        "summary": "Proyectos por ID de usuario",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/idUsuarioPath"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Proyecto"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/usuarios": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Todos los usuarios (filtros)",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SoloEditores": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "activo",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "tipo",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "EDITOR",
                "VISITANTE"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Usuario"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Crear usuario",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Usuario"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/usuarios/{id}": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Usuario por UUID",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SoloEditores": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/idUsuarioPath"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Usuario"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Actualizar usuario",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/idUsuarioPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Usuario"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Eliminar usuario",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SoloEditores": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/idUsuarioPath"
          }
        ],
        "responses": {
          "200": {
            "description": "Eliminado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Usuario"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/usuarios/{id}/baja": {
      "patch": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Dar de baja (activo=false)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/idUsuarioPath"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Usuario"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/usuarios/{id}/alta": {
      "patch": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Dar de alta (activo=true)",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SoloEditores": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/idUsuarioPath"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Usuario"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/usuarios/{id}/password": {
      "patch": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Cambiar contrasena",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SoloEditores": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/idUsuarioPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "password_actual",
                  "password_nueva"
                ],
                "properties": {
                  "password_actual": {
                    "type": "string"
                  },
                  "password_nueva": {
                    "type": "string",
                    "minLength": 8
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Usuario"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/usuarios/estadisticas": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Estadisticas de usuarios",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SoloEditores": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenido de POST /auth/login"
      },
      "SoloEditores": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Mismo token JWT que BearerAuth. El usuario debe tener rol EDITOR en el payload del token."
      }
    },
...
    "parameters": {
      "idUsuarioPath": {
        "name": "id_usuario",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "UUID del usuario"
      }
    }
  }
}